Force fetching PHD2 status after connection
