--- backend/commands.py
@@ -1,8 +1,0 @@
-class StopService:
-    def __call__(self, process=None):
-        process.stop()
-
-
-class GetState:
-    def __call__(self, process=None):
-        process.reply(process.state)
